requirements.txt
numpy==1.20.1
tensorflow==1.15.0
faiss-cpu==1.6.3
absl-py
